reorganized and split base entities
