+++ pype/settings/entities/item_entities.py
@@ -7,207 +7,7 @@ from .lib import (
     SchemeGroupHierarchyBug
 )
 
-from .base_entity import BaseEntity
-
-"""
-# TODO
-Methods:
-- save
-
-# Properties
-## Value attributes - should be set on `set_override_state` based on updated
-    values
-has_defaults
-has_studio_override
-had_studio_override
-has_project_override
-had_project_override
-
-# Abstract properties:
-## Value of an item no matter if has any overrides (without metadata)
-value
-
-## Schema types from schemas
-schema_types
-
-## Unsaved changes
-has_unsaved_changes
-child_is_modified
-
-child_has_studio_override
-child_has_project_override
-
-# Abstract methods:
-## Trigger update of current values(discard changes) and reseting modifications
-set_override_state
-
-## Change current value and trigger modifications and validations of values
-- is not for internal value update
-set
-
-## Return value which will be stored to overrides with metadata
-settings_value
-
-## Update values of defaults, studio overrides and project overrides
-- these should be used anytime to update values
-- current value should be updated if item is without modifications
-update_default_value
-update_studio_values
-update_project_values
-
-## Item should be able register change callbacks
-on_change
-
-## children should notify parent that something has changed
-- value has changed
-- was set to overriden
-- current value has changed
-- etc.
-on_child_change
-
-## Save settings
-save
-
-## Action calls - last to implement
-discard_changes
-add_to_studio_default
-remove_from_studio_default
-remove_from_project_override
-add_to_project_override
-"""
-
-
-class ItemEntity(BaseEntity):
-    def __init__(self, schema_data, parent, is_dynamic_item=False):
-        super(ItemEntity, self).__init__(schema_data, parent, is_dynamic_item)
-
-        self.create_schema_object = self.parent.create_schema_object
-
-        self.is_file = schema_data.get("is_file", False)
-        self.is_group = schema_data.get("is_group", False)
-        self.is_in_dynamic_item = bool(
-            not is_dynamic_item
-            and (parent.is_dynamic_item or parent.is_in_dynamic_item)
-        )
-
-        # Root item reference
-        self.root_item = self.parent.root_item
-
-        # File item reference
-        if self.parent.is_file:
-            self.file_item = self.parent
-        elif self.parent.file_item:
-            self.file_item = self.parent.file_item
-
-        # Group item reference
-        if self.parent.is_group:
-            self.group_item = self.parent
-        elif self.parent.group_item:
-            self.group_item = self.parent.group_item
-
-        # Dynamic item can't have key defined in it-self
-        # - key is defined by it's parent
-        if self.is_dynamic_item:
-            self.require_key = False
-
-        # If value should be stored to environments
-        self.env_group_key = schema_data.get("env_group_key")
-        self.is_env_group = bool(self.env_group_key is not None)
-
-        roles = schema_data.get("roles")
-        if roles is None:
-            roles = parent.roles
-        elif not isinstance(roles, list):
-            roles = [roles]
-        self.roles = roles
-
-        # States of inputs
-        # QUESTION has usage in entity?
-        self.state = None
-
-        self.key = schema_data.get("key")
-        self.label = schema_data.get("label")
-
-        self.item_initalization()
-
-    def schema_validations(self):
-        if self.valid_value_types is NOT_SET:
-            raise ValueError("Attribute `valid_value_types` is not filled.")
-
-        if self.require_key and not self.key:
-            error_msg = "{}: Missing \"key\" in schema data. {}".format(
-                self.path, str(self.schema_data).replace("'", '"')
-            )
-            raise KeyError(error_msg)
-
-        if not self.label and self.is_group:
-            raise ValueError(
-                "{}: Item is set as `is_group` but has empty `label`.".format(
-                    self.path
-                )
-            )
-
-        if self.is_group and self.group_item:
-            raise SchemeGroupHierarchyBug(self.path)
-
-        if not self.file_item and self.is_env_group:
-            raise ValueError((
-                "{}: Environment item is not inside file"
-                " item so can't store metadata for defaults."
-            ).format(self.path))
-
-        if self.label and self.is_dynamic_item:
-            raise ValueError((
-                "{}: Item has set label but is used as dynamic item."
-            ).format(self.path))
-
-    @abstractmethod
-    def item_initalization(self):
-        pass
-
-    def save(self):
-        """Call save on root item."""
-        self.root_item.save()
-
-
-class GUIEntity(ItemEntity):
-    gui_type = True
-
-    schema_types = ["divider", "splitter", "label"]
-    child_has_studio_override = False
-    has_unsaved_changes = False
-    child_is_modified = False
-    child_has_project_override = False
-    value = NOT_SET
-
-    path = "GUIEntity"
-
-    # Abstract methods
-    get_child_path = None
-    set = None
-    set_override_state = None
-    settings_value = None
-    on_change = None
-    on_child_change = None
-
-    _discard_changes = None
-    _remove_from_studio_default = None
-    _remove_from_project_override = None
-    add_to_project_override = None
-    add_to_studio_default = None
-    update_default_value = None
-    update_studio_values = None
-    update_project_values = None
-
-    def __getitem__(self, key):
-        return self.schema_data[key]
-
-    def schema_validations(self):
-        return
-
-    def item_initalization(self):
-        self.valid_value_types = tuple()
-        self.require_key = False
+from .base_entity import ItemEntity
 
 
 class PathEntity(ItemEntity):
